+++ src/engines/reference_engine_class.F90
@@ -338,7 +338,7 @@ contains
          sad_wf = uhf(fractional_uniform_valence=.true., &
                       multiplicity=multiplicity)
 !
-         call sad_wf%prepare([center],  embedding=.false., charge=0)
+         call sad_wf%prepare([center],  embedding=.false., charge=center%charge)
 !
          sad_solver = scf_solver(restart=.false.,                       &
                                  ao_density_guess=ao_density_guess,     &